+++ src/spintop_openhtf/plugs/base.py
@@ -5,6 +5,14 @@ from spintop_openhtf import conf, plugs
 DEFAULT_FORMAT = '%(asctime)-15s: %(message)s'
 
 class UnboundPlug(object):
+    """ A generic interface base class that allows intelligent creation of OpenHTF plugs
+    without limiting its usage to OpenHTF.
+
+    Attributes:
+        logger:
+            If inside an OpenHTF plug, this is the OpenHTF provided logger. If not, it is 
+            a logger with the object ID as name.
+    """
     def __init__(self):
         super().__init__()
         if not hasattr(self, 'logger'):
@@ -13,21 +21,25 @@ class UnboundPlug(object):
             self.logger = logging.getLogger(str(id(self)))
 
     def tearDown(self):
-        """Tear down the plug instance."""
+        """Tear down the plug instance. This is part of the OpenHTF Plug contract"""
         self.close()
         self.logger.close_log()
 
     def open(self, *args, **kwargs):
+        """Abstract method: Open resources related to this interface."""
         raise NotImplementedError()
 
     def close(self):
+        """Abstract method: Close resources related to this interface."""
         raise NotImplementedError()
 
     def log_to_stream(self, stream=None, **kwargs):
+        """Create and add to :attr:`logger` a StreamHandler that streams to :obj:`stream`"""
         handler = logging.StreamHandler(stream)
         self._add_handler(handler, **kwargs)
 
     def log_to_filename(self, filename, **kwargs):
+        """Create and add to :attr:`logger` a FileHandler that logs to :obj:`filename`."""
         handler = logging.FileHandler(filename)
         self._add_handler(handler, **kwargs)
 
@@ -38,6 +50,7 @@ class UnboundPlug(object):
         self.logger.addHandler(handler)
 
     def close_log(self):
+        """Close all :attr:`logger` handlers."""
         for handler in self.logger.handlers:
             handler.close()
 
@@ -46,21 +59,22 @@ class UnboundPlug(object):
         """ Create a bound plug that will retrieve values from conf or the passed values here.
 
         Take SSHInterface for example. 
+        
+        .. highlight:: python
+        .. code-block:: python
 
-        ```python
-        from spintop_openhtf.plugs import from_conf, SSHInterface
+            from spintop_openhtf.plugs import from_conf
+            from spintop_openhtf.plugs.ssh import SSHInterface
 
-        MySSHInterface = SSHInterface.as_plug(
-            'MySSHInterface', # The name of this plug as it will appear in logs
-            addr=from_conf( # from_conf will retrieve the conf value named like this.
-                'my_ssh_addr',
-                description="The addr of my device."
-            ), 
-            username='x', # Always the same
-            password='y'  # Always the same
-        )
-
-        ```
+            MySSHInterface = SSHInterface.as_plug(
+                'MySSHInterface', # The name of this plug as it will appear in logs
+                addr=from_conf( # from_conf will retrieve the conf value named like this.
+                    'my_ssh_addr',
+                    description="The addr of my device."
+                ), 
+                username='x', # Always the same
+                password='y'  # Always the same
+            )
         """
         return plug_factory(cls, name, **kwargs_values)